Move init_api_experimental_auth to init_security
