+++ airflow/www/app.py
@@ -24,44 +24,23 @@ from flask_appbuilder import SQLA
 from flask_caching import Cache
 from flask_wtf.csrf import CSRFProtect
 
-from airflow import settings, version
+from airflow import settings
 from airflow.configuration import conf
 from airflow.logging_config import configure_logging
 from airflow.utils.json import AirflowJsonEncoder
 from airflow.www.extensions.init_appbuilder import init_appbuilder
+from airflow.www.extensions.init_appbuilder_links import init_appbuilder_links
 from airflow.www.extensions.init_jinja_globals import init_jinja_globals
+from airflow.www.extensions.init_manifest_files import configure_manifest_files
+from airflow.www.extensions.init_security import init_api_experimental_auth, init_xframe_protection
 from airflow.www.extensions.init_session import init_logout_timeout, init_permanent_session
 from airflow.www.extensions.init_views import (
     init_api_connexion, init_api_experimental, init_appbuilder_views, init_error_handlers, init_flash_views,
     init_plugins,
 )
 from airflow.www.extensions.init_wsgi_middlewares import init_wsg_middleware
-from airflow.www.static_config import configure_manifest_files
 
 app: Optional[Flask] = None
-
-
-def init_appbuilder_links(app):
-    appbuilder = app.appbuilder
-    if "dev" in version.version:
-        airflow_doc_site = "https://airflow.readthedocs.io/en/latest"
-    else:
-        airflow_doc_site = 'https://airflow.apache.org/docs/{}'.format(version.version)
-
-    appbuilder.add_link("Website",
-                        href='https://airflow.apache.org',
-                        category="Docs",
-                        category_icon="fa-globe")
-    appbuilder.add_link("Documentation",
-                        href=airflow_doc_site,
-                        category="Docs",
-                        category_icon="fa-cube")
-    appbuilder.add_link("GitHub",
-                        href='https://github.com/apache/airflow',
-                        category="Docs")
-    appbuilder.add_link("REST API Reference",
-                        href='/api/v1./api/v1_swagger_ui_index',
-                        category="Docs")
 
 
 def sync_appbuilder_roles(app):
@@ -104,11 +83,7 @@ def create_app(config=None, testing=False, app_name="Airflow"):
     db.session = settings.Session
     db.init_app(app)
 
-    def init_api_experimental_auth():
-        from airflow import api
-        api.load_auth()
-        api.API_AUTH.api_auth.init_app(app)
-    init_api_experimental_auth()
+    init_api_experimental_auth(app)
 
     Cache(app=app, config={'CACHE_TYPE': 'filesystem', 'CACHE_DIR': '/tmp'})
 
@@ -118,16 +93,6 @@ def create_app(config=None, testing=False, app_name="Airflow"):
     configure_manifest_files(app)
 
     with app.app_context():
-
-
-        def init_xframe_protection(app):
-            @app.after_request
-            def apply_caching(response):
-                _x_frame_enabled = conf.getboolean('webserver', 'X_FRAME_ENABLED', fallback=True)
-                if not _x_frame_enabled:
-                    response.headers["X-Frame-Options"] = "DENY"
-                return response
-
         init_appbuilder(app)
 
         init_appbuilder_views(app)
