+++ airflow/www/extensions/init_manifest_files.py
@@ -37,8 +37,9 @@ def configure_manifest_files(app):
         # noinspection PyBroadException
         try:
             global manifest
-            manifest_file = os.path.join(os.path.dirname(__file__),
-                                         'static/dist/manifest.json')
+            manifest_file = os.path.join(
+                os.path.dirname(__file__), os.pardir, 'static/dist/manifest.json'
+            )
             with open(manifest_file, 'r') as file:
                 manifest.update(json.load(file))
